pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Hexlet Page Analyzer — Flask app"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "Flask>=3.0.0,<4.0.0",
  "gunicorn>=21.2.0,<22.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
]

[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["page_analyzer"]

[tool.ruff]
line-length = 100
target-version = "py310"
lint.select = [
  "E", "F", "W",  # pycodestyle/pyflakes/warnings
  "I",              # isort
]

[tool.uv]
dev-dependencies = [
  "ruff>=0.5.0,<1.0.0",
  "pytest>=8.2.0,<9.0.0",
]
